requirements.txt
boto3==1.13.*
django==2.2.*
djangorestframework==3.11.*
envs==1.3.*
requests==2.23.*

# testing
pytest-django==3.9.*
pytest==5.4.*
